requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python
tensorflow
pillow